🚧 Begin transition to new property model
Instead of using a class hierarchy to define properties, define properties as simple data structures that know how to get/set values for their owner type.  Then wrap those properties using callbacks to convert from Control-types to the corresponding HtmlElement

Also, begin separation of editors from the properties - for both better reuse of common editors (text, etc.) and in anticipation of binding in the future - where the human-editable editors won't be available.

Inspired the earlier attempt at this, which can be found in PR #16