Add EnumPropertyEditor as a DropDown for the font-family
In order to be able to select a font-family, implement a Select control for all enums that have a restricted value set.

This necessitated a refactoring of proxied-properties so that all property-accesses (like those for enums) would go through to the original property, instead of each one needing to be individually implemented in the DelegatedProperty.  This also makes it appear like the original property is presented to editors - with the exception that they target Controls instead of HtmlElements.  We may need to revisit this decision in the future if we decide that the Proxy implementation is too slow (haven't checked) or if nobody actually cares if it acts like the original instance or not.